Minor refactoring to make types match.

+++ include/xlsxwriter/worksheet.h
@@ -1646,7 +1646,7 @@ typedef struct lxw_worksheet {
     char *quoted_name;
     char *tmpdir;
 
-    uint32_t index;
+    uint16_t index;
     uint8_t active;
     uint8_t selected;
     uint8_t hidden;
